requirements.txt
chainer==2.0.1
six==1.10.0
dill==0.2.7
PyYAML==3.12
tqdm==4.14.0  # needed for train.py and test.py only                                                     
mlconf==0.0.2 # needed for train.py and test.py only
pynput==1.3.5 # needed for visualisation                                                      
